+++ .gitignore
@@ -34,3 +34,4 @@ yarn-error.log*
 *.tsbuildinfo
 next-env.d.ts
 .env
+pnpm-lock.yaml
