+++ src/app/admin/page.tsx
@@ -1,0 +1,1 @@
+// Admin panel
