+++ src/app/api/projects/[id]/route.ts
@@ -1,0 +1,1 @@
+// /api/projects/[id] route
