+++ src/app/api/projects/route.ts
@@ -1,0 +1,2 @@
+// /api/projects route
+// Should only be used for post requests to create a new project
