+++ src/app/api/user/[id]/route.ts
@@ -1,0 +1,1 @@
+// /api/user/[id] route
