+++ src/app/api/user/route.ts
@@ -1,17 +1,19 @@
-// Sample NextJS 13 API route to return a user object
+// /api/user route
+// Should only be used for post requests to create a new user
+// All other requests should be made to /api/user/[id]
 import { NextResponse } from "next/server";
 
-export async function GET(request: Request) {
-  const sampleUser = {
-    name: "John Doe",
-  };
-  return NextResponse.json({ user: sampleUser }, { status: 200 });
-}
+// export async function GET(request: Request) {
+//   const sampleUser = {
+//     name: "John Doe",
+//   };
+//   return NextResponse.json({ user: sampleUser }, { status: 200 });
+// }
 
-export async function POST(request: Request) {
-  const sampleUser = {
-    name: "John Doe",
-  };
-  console.log(request.body);
-  return NextResponse.json({ user: sampleUser }, { status: 200 });
-}
+// export async function POST(request: Request) {
+//   const sampleUser = {
+//     name: "John Doe",
+//   };
+//   console.log(request.body);
+//   return NextResponse.json({ user: sampleUser }, { status: 200 });
+// }
