+++ src/app/profile/[user]/edit/page.tsx
@@ -1,0 +1,1 @@
+// Edit profile page
