+++ src/app/profile/[user]/page.tsx
@@ -1,0 +1,11 @@
+// The page in which you view a user's profile.
+import React from "react";
+
+export default function Page() {
+  return (
+    <div>
+      Page
+      <h1>hi</h1>
+    </div>
+  );
+}
