+++ src/app/roster/page.tsx
@@ -1,0 +1,2 @@
+// Roster view page
+// Query all users, and then filter them with dropdowns on the client side.
